fix issues found with clang-tidy

+++ Source/Particles/ElementaryProcess/QEDInternals/BreitWheelerEngineWrapper.cpp
@@ -72,7 +72,7 @@ BreitWheelerEngine::init_lookup_tables_from_raw_data (
     if(size_first <= 0 || size_first >= raw_data.size() ) return false;
 
     const auto raw_dndt_table = vector<char>{
-        raw_iter, raw_iter+size_first};
+        raw_iter, raw_iter+static_cast<long>(size_first)};
 
     const auto raw_pair_prod_table = vector<char>{
         raw_iter+static_cast<long>(size_first), raw_data.end()};
